Finalización

+++ src/main/java/com/christianrondon/ecommerce/dto/cartitem/response/CartItemResponse.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
@@ -15,6 +13,6 @@ public class CartItemResponse {
     private String productName;
     private double productPrice;
     private int quantity;
-    private BigDecimal subtotal;
+    private double subtotal;
 
 }
